Make order_indexed_files return dcp error codes.
Calling order_indexed_files never invokes the log.  It is the
responsibility of a higher layer (in opendcp_j2k) to inspect
the return code from order_indexed_files, possibly log an error,
and determine what course of action to take.  Maybe change in
the future, since the error codes aren't terribly descriptive.
For example, they can't tell you *which* file is behaving badly,
only that some file is behaving badly.  Users will probably be
interested in diagnosing the problem as quickly as possible.